Moving state and layout around
