--- src/Footer.js
@@ -1,17 +1,0 @@
-/** @jsxImportSource @emotion/react */
-
-import { Navigation } from './Navigation';
-
-export function Footer() {
-	return (
-		<footer
-			css={{
-				padding: '1rem 0',
-				borderTop: '2px dashed var(--border)',
-				textAlign: 'center',
-			}}
-		>
-			<Navigation />
-		</footer>
-	);
-}
